backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
pydantic-settings==2.5.2
celery[redis]==5.4.0
redis==5.1.0
smbprotocol==1.13.0
faster-whisper==1.0.3
openai==1.52.0
httpx==0.27.0
cryptography==43.0.3
python-dotenv==1.0.1
pysrt==1.1.2
croniter==3.0.3
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.30.0
